+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## TO BE RELEASED
 
+* Create a squid3 proxy to be used by the zadara storage array to get to s3 for
+  object store backups.  See
+  [mh-opsworks](https://github.com/harvard-dce/mh-opsworks/blob/master/README.zadara.md)
+  for more information about how to use this recipe. This recipe is independent
+  of deployments.
+
 ## 1.1.5 - 2/25/2016
 
 * *REQUIRES MANUAL CHEF RECIPE RUNS:* Set the bash prompt to be in color and